chore: move EmbraceInstrumentationBase to instrumentations/
