+++ src/instrumentations/exceptions/GlobalExceptionInstrumentation/types.ts
@@ -1,4 +1,4 @@
-import type { EmbraceInstrumentationBaseArgs } from '../../session/index.js';
+import type { EmbraceInstrumentationBaseArgs } from '../../EmbraceInstrumentationBase/index.js';
 
 export type GlobalExceptionInstrumentationArgs = Pick<
   EmbraceInstrumentationBaseArgs,
