+++ src/instrumentations/session/SpanSessionBrowserActivityInstrumentation/types.ts
@@ -1,4 +1,4 @@
-import type { EmbraceInstrumentationBaseArgs } from '../EmbraceInstrumentationBase/types.js';
+import type { EmbraceInstrumentationBaseArgs } from '../../EmbraceInstrumentationBase/types.js';
 
 export type SpanSessionBrowserActivityInstrumentationArgs = Pick<
   EmbraceInstrumentationBaseArgs,
