+++ src/instrumentations/session/SpanSessionTimeoutInstrumentation/types.ts
@@ -1,4 +1,4 @@
-import type { EmbraceInstrumentationBaseArgs } from '../EmbraceInstrumentationBase/types.js';
+import type { EmbraceInstrumentationBaseArgs } from '../../EmbraceInstrumentationBase/types.js';
 
 export type SpanSessionTimeoutInstrumentationArgs = Pick<
   EmbraceInstrumentationBaseArgs,
