+++ src/instrumentations/session/SpanSessionVisibilityInstrumentation/types.ts
@@ -1,4 +1,4 @@
-import type { EmbraceInstrumentationBaseArgs } from '../EmbraceInstrumentationBase/types.js';
+import type { EmbraceInstrumentationBaseArgs } from '../../EmbraceInstrumentationBase/types.js';
 
 export type SpanSessionVisibilityInstrumentationArgs = Pick<
   EmbraceInstrumentationBaseArgs,
