+++ src/instrumentations/session/index.ts
@@ -3,6 +3,6 @@ export { SpanSessionVisibilityInstrumentation } from './SpanSessionVisibilityIns
 export { SpanSessionBrowserActivityInstrumentation } from './SpanSessionBrowserActivityInstrumentation/index.js';
 export { SpanSessionOnLoadInstrumentation } from './SpanSessionOnLoadInstrumentation/index.js';
 export { SpanSessionTimeoutInstrumentation } from './SpanSessionTimeoutInstrumentation/index.js';
-export type { EmbraceInstrumentationBaseArgs } from './EmbraceInstrumentationBase/index.js';
-export { EmbraceInstrumentationBase } from './EmbraceInstrumentationBase/index.js';
+export type { EmbraceInstrumentationBaseArgs } from '../EmbraceInstrumentationBase/index.js';
+export { EmbraceInstrumentationBase } from '../EmbraceInstrumentationBase/index.js';
 export type { SessionSpan } from './types.js';
